Fixed up the schema/doepages harvester to use the new system, as an example
